backend/requirements.txt
Flask==3.0.0
python-dotenv==1.0.1
requests==2.31.0
cachetools==5.3.3
google-cloud-secret-manager==2.19.0